+++ src/Types/IsbnType.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Tpetry\PostgresqlEnhanced\Types;
 
-class XmlType extends BaseType
+class IsbnType extends BaseType
 {
     /**
      * Gets the name of this type.
      */
     public function getName()
     {
-        return 'xml';
+        return 'isbn';
     }
 }